% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_brand_occurrences.R
\name{count_brand_occurrences}
\alias{count_brand_occurrences}
\title{Counts the number of cars per brand}
\usage{
count_brand_occurrences(df)
}
\arguments{
\item{df}{A data frame with a column \code{model_name}.}
}
\value{
A named integer vector with labels as names and number of occurrences.
}
\description{
This function counts how often vehicles of a specific car brand
appear in a list of model names. It checks
whether the model name matches an entry from the car_brands list.
}
\examples{
data(example_hotwheels)
df <- example_hotwheels
df <- standardize_column_names(df)
count_brand_occurrences(df)
}
