% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_color_cars.R
\name{count_color_cars}
\alias{count_color_cars}
\title{gives the number of cars with a specific colure}
\usage{
count_color_cars(df, color)
}
\arguments{
\item{df}{data set with hotwheels from the side: https://collecthw.com/}

\item{color}{A string representing the color to count (case-insensitive).}
}
\value{
The nummber of cars for a specific colure
}
\description{
gives the number of cars with a specific colure.
}
\examples{
df <- standardize_column_names(example_hotwheels)
count_color_cars(df, "red")
}
