% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_hotwheels.R
\docType{data}
\name{example_hotwheels}
\alias{example_hotwheels}
\title{Example Hot Wheels Data}
\format{
A data frame from the webside https://collecthw.com/:
\describe{
\item{Link}{A URL or reference link to more information about the car}
\item{Amount.Owned}{The number of that particular model owned by the collector.}
\item{Year}{The release year of the toy car.}
\item{Collector}{Number of the car in the series}
\item{Series #}{The ID of the series}
\item{Model.Name}{The specific model name of the car (e.g., Mustang GT).}
\item{Series}{The name of the series}
\item{Color}{The primary color of the toy car.}
\item{...}{}
}
}
\source{
https://collecthw.com/
}
\usage{
example_hotwheels
}
\description{
A sample dataset for demonstrating functions.
}
\keyword{datasets}
