% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_unique_models.R
\name{list_unique_models}
\alias{list_unique_models}
\title{List of all different models in the Hot Wheels collection}
\usage{
list_unique_models(df)
}
\arguments{
\item{df}{Hot Wheels data frame}
}
\value{
A character vector with all unique model names
}
\description{
Returns an alphabetically sorted list of all unique model names.
}
\examples{
data(example_hotwheels)
df <- standardize_column_names(example_hotwheels)
list_unique_models(df)

}
