% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_unique_series.R
\name{list_unique_series}
\alias{list_unique_series}
\title{List of all different series in the Hot Wheels collection}
\usage{
list_unique_series(df)
}
\arguments{
\item{df}{Hot Wheels data frame}
}
\value{
A character vector with all unique series names
}
\description{
Returns a list of all different series names.
}
\examples{
data(example_hotwheels)
df <- standardize_column_names(example_hotwheels)
list_unique_series(df)

}
