% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_column_names.R
\name{standardize_column_names}
\alias{standardize_column_names}
\title{Standardize Column Names}
\usage{
standardize_column_names(df)
}
\arguments{
\item{df}{A data frame whose column names should be standardized.}
}
\value{
The same data frame with cleaned column names.
}
\description{
Converts column names to lower case and replaces spaces with underscores.
}
\examples{
data(example_hotwheels)
df <- standardize_column_names(example_hotwheels)
}
